Handle shallow clone problems
