+++ application/common/tests/test_dynamodb.py
@@ -2,78 +2,12 @@ from decimal import Decimal
 from json import dumps, loads
 from os import environ
 from time import time
-from typing import Any
 from unittest.mock import patch
 
-import pytest
 from aws_lambda_powertools.logging import Logger
 from boto3.dynamodb.types import TypeDeserializer
 
 FILE_PATH = "application.common.dynamodb"
-
-
-@pytest.fixture()
-def dynamodb_table_create(dynamodb_client) -> dict[str, Any]:
-    """Create a DynamoDB CHANGE_EVENTS_TABLE table pytest.fixture."""
-    return dynamodb_client.create_table(
-        TableName=environ["CHANGE_EVENTS_TABLE_NAME"],
-        BillingMode="PAY_PER_REQUEST",
-        KeySchema=[
-            {"AttributeName": "Id", "KeyType": "HASH"},
-            {"AttributeName": "ODSCode", "KeyType": "RANGE"},
-        ],
-        AttributeDefinitions=[
-            {"AttributeName": "Id", "AttributeType": "S"},
-            {"AttributeName": "ODSCode", "AttributeType": "S"},
-            {"AttributeName": "SequenceNumber", "AttributeType": "N"},
-        ],
-        GlobalSecondaryIndexes=[
-            {
-                "IndexName": "gsi_ods_sequence",
-                "KeySchema": [
-                    {"AttributeName": "ODSCode", "KeyType": "HASH"},
-                    {"AttributeName": "SequenceNumber", "KeyType": "RANGE"},
-                ],
-                "Projection": {"ProjectionType": "ALL"},
-            },
-        ],
-    )
-
-
-def test_get_circuit_is_open_none(dynamodb_table_create, dynamodb_client):
-    from application.common.dynamodb import get_circuit_is_open
-
-    assert get_circuit_is_open("BLABLABLA") is None
-
-
-def test_put_and_get_circuit_is_open(dynamodb_table_create, dynamodb_client):
-    from application.common.dynamodb import get_circuit_is_open, put_circuit_is_open
-
-    put_circuit_is_open("TESTCIRCUIT", True)
-
-    assert get_circuit_is_open("TESTCIRCUIT")
-
-
-def test_put_circuit_exception(dynamodb_table_create, dynamodb_client):
-    from application.common.dynamodb import put_circuit_is_open
-
-    temp_table = environ["CHANGE_EVENTS_TABLE_NAME"]
-    del environ["CHANGE_EVENTS_TABLE_NAME"]
-    with pytest.raises(Exception):  # noqa: PT011,B017
-        put_circuit_is_open("TESTCIRCUIT", True)
-
-    environ["CHANGE_EVENTS_TABLE_NAME"] = temp_table
-
-
-def test_get_circuit_exception(dynamodb_table_create, dynamodb_client):
-    from application.common.dynamodb import get_circuit_is_open
-
-    temp_table = environ["CHANGE_EVENTS_TABLE_NAME"]
-    del environ["CHANGE_EVENTS_TABLE_NAME"]
-    with pytest.raises(Exception):  # noqa: PT011,B017
-        get_circuit_is_open("TESTCIRCUIT")
-
-    environ["CHANGE_EVENTS_TABLE_NAME"] = temp_table
 
 
 def test_add_change_event_to_dynamodb(dynamodb_table_create, change_event, dynamodb_client):
@@ -249,37 +183,3 @@ def copy_and_modify_website(ce, new_website: str):
     copy = ce.copy()
     copy["Contacts"][0]["ContactValue"] = new_website
     return copy
-
-
-def test_get_newest_event_per_odscode(dynamodb_table_create, change_event, dynamodb_client, dynamodb_resource):
-    from application.common.dynamodb import add_change_event_to_dynamodb, get_newest_event_per_odscode
-
-    ceAAA11 = change_event.copy()  # noqa: N806
-    ceAAA11["ODSCode"] = "AAA11"
-    add_change_event_to_dynamodb(ceAAA11, 301, int(time()))
-    for i in range(10):
-        add_change_event_to_dynamodb(ceAAA11, i, int(time()))
-
-    ceBBB22 = change_event.copy()  # noqa: N806
-    ceBBB22["ODSCode"] = "BBB22"
-    add_change_event_to_dynamodb(ceBBB22, 505, int(time()))
-    for i in range(10):
-        add_change_event_to_dynamodb(ceBBB22, i, int(time()))
-
-    ceCCC33 = change_event.copy()  # noqa: N806
-    ceCCC33["ODSCode"] = "CCC33"
-    add_change_event_to_dynamodb(ceCCC33, 400, int(time()))
-    for i in range(10):
-        add_change_event_to_dynamodb(ceCCC33, i, int(time()))
-
-    expected_seq_num = {"AAA11": 301, "BBB22": 505, "CCC33": 400}
-
-    resp = get_newest_event_per_odscode()
-    assert len(resp) == 3
-    for resp_event in resp.values():
-        assert resp_event["SequenceNumber"] == expected_seq_num[resp_event["ODSCode"]]
-
-    resp = get_newest_event_per_odscode(limit=1)
-    assert len(resp) == 3
-    for resp_event in resp.values():
-        assert resp_event["SequenceNumber"] == expected_seq_num[resp_event["ODSCode"]]
